requirements_mnist.txt
streamlit==1.29.0
scikit-learn==1.3.2
numpy==1.24.3
matplotlib==3.7.1
Pillow==10.1.0
opencv-python-headless==4.8.1.78
streamlit-drawable-canvas==0.9.3
joblib==1.3.2